Renombrado, traslado de extensiones de assertpy y fixtures
Renombrado archivo de 'importar.py' a 'conftest.py' siguiendo las prácticas recomendadas por Pytest.
Extensiones de assertpy implementadas movidas a este archivo.
Fixtures utilizados en los tests movidos a este archivo.
