+++ test/conftest.py
@@ -1,7 +1,11 @@
+import pytest
 from opotests.dificultad import Dificultad
 from opotests.grupo import Grupo
 from opotests.tema import Tema
+from opotests.pregunta import Pregunta
+from opotests.oposicion import Oposicion
 from datetime import date, datetime
+from assertpy import add_extension
 
 def importar_pregunta(ruta):
     """
@@ -88,3 +92,46 @@ def grupo(str):
                'E': Grupo.E}
     return grupos[str]
     
+@pytest.fixture
+def pregunta():
+    datos_pregunta = importar_pregunta('test/preguntas.txt')
+    return Pregunta(datos_pregunta[0], datos_pregunta[1], datos_pregunta[2], datos_pregunta[3], datos_pregunta[4], datos_pregunta[5])
+
+@pytest.fixture
+def oposicion():
+    datos_oposicion= importar_oposicion('test/oposiciones.txt')
+    return Oposicion(datos_oposicion[0], datos_oposicion[1], datos_oposicion[2], datos_oposicion[3])
+
+def esta_entre_los_temas(self, lista):
+    """
+    Extension de assertpy para verificar si un tema (titulo) esta entre los temas de una lista
+    """
+    esta = False
+    for tema in lista:
+        if self.val == tema.titulo:
+            esta = True
+            break
+    if esta == False:
+        self.error(f'{self.val} no esta')
+
+    return self
+
+# Añadimos la extension
+add_extension(esta_entre_los_temas)
+
+def esta_entre_las_claves(self, diccionario):
+    """
+    Extension de assertpy para verificar si un valor esta entre las claves de un diccionario
+    """
+    esta = False
+    for key in diccionario.keys():
+        if self.val == key:
+            esta = True
+            break
+    if esta == False:
+        self.error(f'{self.val} no esta')
+
+    return self
+
+# Añadimos la extension
+add_extension(esta_entre_las_claves)
