+++ test/test_oposiciones.py
@@ -1,21 +1,21 @@
 import pytest
-from assertpy import assert_that, add_extension
+from assertpy import assert_that
 from datetime import date
 
-from importar import *
+from conftest import *
 from opotests.tema import Tema
 from opotests.oposicion import Oposicion
 from opotests.pregunta import Pregunta
 
-@pytest.fixture
-def pregunta():
-    datos_pregunta = importar_pregunta('test/preguntas.txt')
-    return Pregunta(datos_pregunta[0], datos_pregunta[1], datos_pregunta[2], datos_pregunta[3], datos_pregunta[4], datos_pregunta[5])
+# @pytest.fixture
+# def pregunta():
+#     datos_pregunta = importar_pregunta('test/preguntas.txt')
+#     return Pregunta(datos_pregunta[0], datos_pregunta[1], datos_pregunta[2], datos_pregunta[3], datos_pregunta[4], datos_pregunta[5])
 
-@pytest.fixture
-def oposicion():
-    datos_oposicion= importar_oposicion('test/oposiciones.txt')
-    return Oposicion(datos_oposicion[0], datos_oposicion[1], datos_oposicion[2], datos_oposicion[3])
+# @pytest.fixture
+# def oposicion():
+#     datos_oposicion= importar_oposicion('test/oposiciones.txt')
+#     return Oposicion(datos_oposicion[0], datos_oposicion[1], datos_oposicion[2], datos_oposicion[3])
 
 def test_estado_oposicion():
     """
@@ -64,20 +64,4 @@ def test_pregunta_tema_oposicion():
     oposicion = Oposicion(datos_oposicion[0], datos_oposicion[1], datos_oposicion[2], datos_oposicion[3])
     
     assert_that(pregunta.tema).esta_entre_los_temas(oposicion.temas)
-
-def esta_entre_los_temas(self, lista):
-    """
-    Extension de assertpy para verificar si un tema (titulo) esta entre los temas de una lista
-    """
-    esta = False
-    for tema in lista:
-        if self.val == tema.titulo:
-            esta = True
-            break
-    if esta == False:
-        self.error(f'{self.val} no esta')
-
-    return self
-
-# Añadimos la extension
-add_extension(esta_entre_los_temas)
+
